updating files

+++ lib/roi_data_layer/roibatchLoader.py
@@ -51,7 +51,7 @@ class roibatchLoader(data.Dataset):
             # for ratio cross 1, we make it to be 1.
             target_ratio = 1
 
-        self.ratio_list_batch[left_idx:(right_idx+1)] = np.float(target_ratio)
+        self.ratio_list_batch[left_idx:(right_idx+1)] = target_ratio
 
 
   def __getitem__(self, index):
